requirements.txt
requests>=2.31.0
colorama>=0.4.6
beautifulsoup4>=4.12.2
pillow>=10.0.0
qrcode>=8.2.0 